requirements.txt
# Core AI orchestration framework
crewai==0.80.0
crewai-tools==0.17.0

# LLM providers
google-generativeai==0.8.3
langchain-google-genai==2.0.7

# Web search
tavily-python==0.5.0

# Financial data
yfinance==0.2.50

# Vector store & RAG
chromadb==0.6.3
langchain==0.3.14
langchain-community==0.3.27

# Data processing
pandas==2.2.3
numpy==2.2.1
ta==0.11.0

# Environment & utilities
python-dotenv==1.0.1
pydantic==2.10.5
